+++ lib/src/bindings/common/platform_user_data.dart
@@ -1,0 +1,1 @@
+final platformUserData = <int, Object>{};