+++ lib/src/bindings/view/platform_view.dart
@@ -8,6 +8,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 
 import 'package:yandex_maps_mapkit_lite/src/bindings/common/mapkit_method_channel.dart';
+import 'package:yandex_maps_mapkit_lite/src/runtime/view/gpu_info.dart';
 import 'platform_view_type.dart';
 
 class PlatformView {
@@ -18,17 +19,28 @@ class PlatformView {
   static const viewType = 'flutter_yandex_maps_view_factory';
   static const channel = const MapkitMethodChannel('view');
 
-  static Widget create(bool Function(PlatformView) onCreate,
-      {required PlatformViewType platformViewType,
-      TextDirection? textDirection}) {
+  static Widget create(
+    bool Function(PlatformView) onCreate, {
+    required PlatformViewType platformViewType,
+    required Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers,
+    required PlatformViewHitTestBehavior hitTestBehavior,
+    TextDirection? textDirection,
+  }) {
     if (Platform.isAndroid) {
       return _AndroidView(
-          platformViewType, textDirection, _onViewCreated(onCreate));
+        platformViewType,
+        textDirection,
+        _onViewCreated(onCreate),
+        gestureRecognizers,
+        hitTestBehavior,
+      );
     } else {
       final widget = UiKitView(
         viewType: viewType,
         layoutDirection: textDirection,
         onPlatformViewCreated: _onViewCreated(onCreate),
+        gestureRecognizers: gestureRecognizers,
+        hitTestBehavior: hitTestBehavior,
       );
 
       return widget;
@@ -40,8 +52,11 @@ class _AndroidView extends StatelessWidget {
   final PlatformViewType platformViewType;
   final TextDirection? textDirection;
   final void Function(int) onCreate;
+  final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+  final PlatformViewHitTestBehavior hitTestBehavior;
 
-  _AndroidView(this.platformViewType, this.textDirection, this.onCreate);
+  _AndroidView(this.platformViewType, this.textDirection, this.onCreate,
+      this.gestureRecognizers, this.hitTestBehavior);
 
   @override
   Widget build(BuildContext context) {
@@ -52,61 +67,33 @@ class _AndroidView extends StatelessWidget {
     final creationParams = <String, String>{};
     switch (platformViewType) {
       case PlatformViewType.Hybrid:
-        _setSurfaceViewType(creationParams);
-        return _createPlatformViewLink((params) =>
-            PlatformViewsService.initExpensiveAndroidView(
-              creationParams: creationParams,
-              creationParamsCodec: const StandardMessageCodec(),
-              id: params.id,
-              viewType: PlatformView.viewType,
-              layoutDirection: textDirection,
-              onFocus: () {
-                params.onFocusChanged(true);
-              },
-            )
-              ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
-              ..addOnPlatformViewCreatedListener(onCreate)
-              ..create());
+        return _createHybridView(creationParams, textDirection);
 
       case PlatformViewType.Virtual:
-        _setSurfaceViewType(creationParams);
-        return AndroidView(
-          viewType: PlatformView.viewType,
-          creationParams: creationParams,
-          creationParamsCodec: const StandardMessageCodec(),
-          layoutDirection: textDirection,
-          onPlatformViewCreated: onCreate,
-        );
+        return _createVirtualView(creationParams, textDirection);
 
       case PlatformViewType.TextureHybrid:
+        return _createTextureHybridView(creationParams, textDirection);
       case PlatformViewType.Compat:
-        _setTextureViewType(creationParams);
-        return _createPlatformViewLink((params) =>
-            PlatformViewsService.initSurfaceAndroidView(
-              creationParams: creationParams,
-              creationParamsCodec: const StandardMessageCodec(),
-              id: params.id,
-              viewType: PlatformView.viewType,
-              layoutDirection: textDirection,
-              onFocus: () {
-                params.onFocusChanged(true);
-              },
-            )
-              ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
-              ..addOnPlatformViewCreatedListener(onCreate)
-              ..create());
+        final gpuVendor = GpuInfoManager.glGpuVendor();
+        final gpuRenderer = GpuInfoManager.glGpuRenderer();
+        // This GPU may be blocked when rendering a map with a large number of vertices.
+        return gpuVendor == 'PowerVR' && gpuRenderer == 'GE8320'
+            ? _createVirtualView(creationParams, textDirection)
+            : _createTextureHybridView(creationParams, textDirection);
     }
   }
 
-  static PlatformViewLink _createPlatformViewLink(
+  PlatformViewLink _createPlatformViewLink(
       PlatformViewController Function(PlatformViewCreationParams)
           onCreatePlatformView) {
     return PlatformViewLink(
       viewType: PlatformView.viewType,
       surfaceFactory: ((context, controller) => AndroidViewSurface(
-          controller: controller as AndroidViewController,
-          hitTestBehavior: PlatformViewHitTestBehavior.opaque,
-          gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{})),
+            controller: controller as AndroidViewController,
+            hitTestBehavior: hitTestBehavior,
+            gestureRecognizers: gestureRecognizers,
+          )),
       onCreatePlatformView: onCreatePlatformView,
     );
   }
@@ -117,6 +104,58 @@ class _AndroidView extends StatelessWidget {
 
   static void _setSurfaceViewType(Map<String, String> params) {
     params['viewType'] = 'surface';
+  }
+
+  Widget _createHybridView(
+      Map<String, String> creationParams, TextDirection textDirection) {
+    _setSurfaceViewType(creationParams);
+    return _createPlatformViewLink(
+        (params) => PlatformViewsService.initExpensiveAndroidView(
+              creationParams: creationParams,
+              creationParamsCodec: const StandardMessageCodec(),
+              id: params.id,
+              viewType: PlatformView.viewType,
+              layoutDirection: textDirection,
+              onFocus: () {
+                params.onFocusChanged(true);
+              },
+            )
+              ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
+              ..addOnPlatformViewCreatedListener(onCreate)
+              ..create());
+  }
+
+  Widget _createVirtualView(
+      Map<String, String> creationParams, TextDirection textDirection) {
+    _setSurfaceViewType(creationParams);
+    return AndroidView(
+      viewType: PlatformView.viewType,
+      creationParams: creationParams,
+      creationParamsCodec: const StandardMessageCodec(),
+      layoutDirection: textDirection,
+      onPlatformViewCreated: onCreate,
+      gestureRecognizers: gestureRecognizers,
+      hitTestBehavior: hitTestBehavior,
+    );
+  }
+
+  Widget _createTextureHybridView(
+      Map<String, String> creationParams, TextDirection textDirection) {
+    _setTextureViewType(creationParams);
+    return _createPlatformViewLink(
+        (params) => PlatformViewsService.initSurfaceAndroidView(
+              creationParams: creationParams,
+              creationParamsCodec: const StandardMessageCodec(),
+              id: params.id,
+              viewType: PlatformView.viewType,
+              layoutDirection: textDirection,
+              onFocus: () {
+                params.onFocusChanged(true);
+              },
+            )
+              ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
+              ..addOnPlatformViewCreatedListener(onCreate)
+              ..create());
   }
 }