+++ lib/src/bindings/widgets/platform_view_widget.dart
@@ -1,3 +1,6 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:yandex_maps_mapkit_lite/src/bindings/view/platform_view.dart';
 import 'package:yandex_maps_mapkit_lite/src/bindings/view/platform_view_type.dart';
@@ -8,11 +11,15 @@ abstract class PlatformViewWidget extends StatefulWidget {
     required this.onViewCreated,
     this.platformViewType = PlatformViewType.Compat,
     this.textDirection,
+    this.gestureRecognizers = const <Factory<OneSequenceGestureRecognizer>>{},
+    this.hitTestBehavior = PlatformViewHitTestBehavior.opaque,
   });
 
   final bool Function(PlatformView) onViewCreated;
   final PlatformViewType platformViewType;
   final TextDirection? textDirection;
+  final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+  final PlatformViewHitTestBehavior hitTestBehavior;
 
   @override
   State createState() => _PlatformViewWidgetState();
@@ -30,6 +37,8 @@ class _PlatformViewWidgetState extends State<PlatformViewWidget> {
       },
       platformViewType: widget.platformViewType,
       textDirection: widget.textDirection,
+      gestureRecognizers: widget.gestureRecognizers,
+      hitTestBehavior: widget.hitTestBehavior,
     );
   }
 }