+++ lib/src/runtime/view/gpu_info_export.dart
@@ -1,0 +1,1 @@
+export 'package:yandex_maps_mapkit_lite/src/runtime/view/gpu_info.dart';